[feats] 넥서스 시스템 - 게임 승리처리 1차 커밋
오류가 발생하여 우선 1차적으로 커밋 후, 오류 수정후 이어갈 예정
